# proto/Test-service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Test-service.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/todo/all": {
      "get": {
        "operationId": "ReadAll",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoReadAllResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "v1",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "TestService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protoReadAllResponse": {
      "type": "object",
      "properties": {
        "v1": {
          "type": "string",
          "format": "int64"
        },
        "toDos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protoToDo"
          }
        }
      }
    },
...
    "protoToDo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        }
      }
    },
...
  }
}